enhance and document termination behavior
The code handling shutdown had a theoretic race condition (one
goroutine creating the CSIDriver object, the other removing it). It's
cleaner to do it without relying on goroutines. Catching all
termination signals makes the code less dependent on Kubernetes
specifics.

Testing this feature showed that an orderly removal of a CSI driver is
non-trivial. This needs to be documented. While at it, the
documentation gets improved (more examples, remove remaining
references to CRD, spelling).